Fix precision problem in hypotenuse()
With this wrong type cast we drop precision along the way, which was
caught with automatic testing.

+++ src/hypotenuse.c
@@ -5,7 +5,7 @@
 #include "util.h"
 
 static double
-hypotenuse(double a, float b)
+hypotenuse(double a, double b)
 {
 	return sqrt(a * a + b * b);
 }
